% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_fun.R
\name{bar_plot}
\alias{bar_plot}
\title{Produces a bar graph using crabs data set}
\usage{
bar_plot(data, column_1, column_2)
}
\arguments{
\item{data}{crabs dataset}

\item{column_1}{x-axis data, independent variable}

\item{column_2}{y-axis data, dependent variable}
}
\value{
bar_plot bar graph of two chosen columns (bar_plot)
}
\description{
Use data set crabs and any two columns to create a bar graph
Returns ggplot with geom_col(), which displays a bar graph
}
