% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{linear_model}
\alias{linear_model}
\title{Produces summary table of a linear model}
\usage{
linear_model(df, dependent, independent_cols)
}
\arguments{
\item{df}{crabs dataset}

\item{dependent}{numeric column used as dependent variable}

\item{independent_cols}{one or more columns used as independent variables}

\item{return}{linear model summary table}
}
\description{
Use data set "crabs" and any numeric columns to build linear model
Returns linear model summary
}
