% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{Produce scatter plot for any two columns of crabs data}
\usage{
plot(data, column_1, column_2)
}
\arguments{
\item{data}{crabs dataset}

\item{column_1}{x-axis data, independent variable}

\item{column_2}{y-axis data, dependent variable}
}
\value{
scatter_plot scatter plot of two chosen columns (plot)
}
\description{
Use the data set "crabs" and any two columns in order to produce a scatter plot.
Returns a ggplot with a geom point, which displays as a graph with scattered data points.
}
