% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{remove_nas}
\alias{remove_nas}
\title{Removes NA values from crabs data set}
\usage{
remove_nas(data, columns)
}
\arguments{
\item{data}{crabs dataset}

\item{columns}{columns from data set}
}
\value{
clean dataset cleaned of NAs
}
\description{
Use the data set "crabs" and any columns to remove NAs from
Returns data set with all NA values omitted
}
